+++ src/test/java/org/scion/jpan/internal/NatMappingTest.java
@@ -26,6 +26,7 @@ import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.scion.jpan.*;
 import org.scion.jpan.testutil.ExamplePacket;
@@ -246,6 +247,7 @@ class NatMappingTest {
     assertEquals("Could not find a STUN/NAT solution for the border router.", e.getMessage());
   }
 
+  @Disabled
   @Test
   void testPrefetch_receive() throws IOException {
     System.setProperty(Constants.PROPERTY_NAT, "BR");